requirements.txt
acres==0.4.1
aiofiles==24.1.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
django-tailwind==4.0.1
djangorestframework==3.16.0
djhtml==3.0.8
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
frontend==0.0.3
h11==0.16.0
httplib2==0.22.0
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isodate==0.6.1
itsdangerous==2.2.0
looseversion==1.3.0
lxml==5.4.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nodeenv==1.9.1
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prov==2.0.1
puremagic==1.29
pydot==4.0.0
Pygments==2.19.1
PyMuPDF==1.26.0
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
requests==2.32.3
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.0
tools==1.0.2
traits==7.0.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
virtualenv==20.31.2
